Beta 4 - Finish tuning, add support for 7th gear
Modified the engine to allow cars to have up to 7 gears instead of just 6 (and scrapped the Drag force value), and finished the tuning menu. Also removed the fade in effect in the intro to attempt to fix a conflict with ASMHOOK, but ASMHOOK still causes the game to restart in races so a warning is now included in the README to disable it.